+++ components/PDFViewer.tsx
@@ -16,8 +16,8 @@ import {
 import { ChartData } from "@/types/chart";
 import { useTheme } from "next-themes";
 import { getFormattedChartName } from "@/lib/chartFormatter";
-import "react-pdf/dist/esm/Page/AnnotationLayer.css";
-import "react-pdf/dist/esm/Page/TextLayer.css";
+import "react-pdf/dist/Page/AnnotationLayer.css";
+import "react-pdf/dist/Page/TextLayer.css";
 
 // Configure worker
 pdfjs.GlobalWorkerOptions.workerSrc = `//unpkg.com/pdfjs-dist@${pdfjs.version}/build/pdf.worker.min.mjs`;